finance-ai/requirements.txt
flask==3.1.0
flask-cors==4.0.0
google-generativeai==0.8.3
prophet==1.1.5
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0
pillow-heif==0.18.0
python-dotenv==1.0.1
pika==1.3.2
gunicorn==22.0.0
pdfplumber==0.11.0
